Improved default double-click behavior...
Instead of opening the "detail view", double-clicking on a log item now immediately takes you straight to the line in the source code where the log statement was made. Using double-click + alt opens the "detail view" instead, since this is a less-likely scenario.

Also, tweaked the AppleScript to make it faster (the call to `open "$file"` was *especially* affecting responsiveness).